Move structs into separate files.
